src/guardian/requirements.txt
rembg==2.0.59
python-dotenv==1.0.1
nats-py==2.9.0
protobuf==5.28.3
grpcio==1.67.0
grpcio-tools==1.67.0
aiohttp==3.10.5